Cosmetic improvements.

+++ teeer/teeer.sh
@@ -20,33 +20,32 @@ die() {
 }
 
 cleanup() {
-    rm -f $ERR_LOG
+    if [ -f "$ERR_LOG" ]; then
+	rm -f $ERR_LOG
+    fi
 }
 
 ### Parse params
-
 ALL_LOG="$1"; shift
 test ! -z "$ALL_LOG" || die "Missing LOGFILE argument."
 test ! -z "$1" || die "Missing COMMAND argument."
 
 ### Set up
-
 set -e
 set -o pipefail
-ERR_LOG="/tmp/$$-err.log"
+ERR_LOG="/tmp/teeer-$$-err.log"
 trap cleanup EXIT
 rm -f $ALL_LOG $ERR_LOG; touch $ERR_LOG
+echo "$ERR_LOG"
 
 ### Subprocess handling stderr
-
 (
     # retrieve error messages from $ERR_LOG, and tee to $ALL_LOG...
     tail -f $ERR_LOG | tee -a $ALL_LOG
 ) >&2 & # ... and to stderr; all this in background
-fd=$! # remember subprocess
+fd=$! # remember subprocess id
 
 ### Command
-
 # run command, stream err to $ERR_LOG, and tee out to stdout and $ALL_LOG
 $* 2>$ERR_LOG | tee -a $ALL_LOG
